requirements.txt
# 🖥️ Streamlit — easy way to build interactive web apps for your AI project
streamlit

# 🔁 Nest Asyncio — allows nested event loops (useful when using async code in notebooks or Streamlit)
nest_asyncio

# 📄 Docling — helps parse and extract text from PDF, DOCX, images, etc.
docling

# 🧠 Qdrant Client — Python SDK to interact with Qdrant (a vector database)
qdrant-client

# 🌱 Python Dotenv — loads environment variables from a .env file (e.g. API keys)
python-dotenv

# 🤖 OpenAI — official SDK for calling OpenAI models and embeddings
openai

# ⏳ TQDM — shows progress bars for loops and background processes
tqdm

# 🧬 Transformers — powerful NLP/multimodal library from Hugging Face
transformers

# 🦙 Llama Index — framework to build Retrieval-Augmented Generation (RAG) pipelines
llama-index
llama_index.embeddings.huggingface

AzureOpenAI
einops
gpt-oss
mlx-vlm